+++ csrc/batch_prefill_sm90.cu
@@ -70,7 +70,7 @@ void BatchPrefillWithRaggedKVCacheSM90Run(
     at::Tensor float_workspace_buffer, at::Tensor int_workspace_buffer,
     at::Tensor plan_info_vec, at::Tensor q, at::Tensor k, at::Tensor v,
     at::Tensor qo_indptr, at::Tensor kv_indptr, at::Tensor o, std::optional<at::Tensor> maybe_lse,
-    int64_t mask_mode_code, int64_t layout, int32_t window_left ADDITIONAL_FUNC_PARAMS,
+    int64_t mask_mode_code, int64_t layout, int64_t window_left ADDITIONAL_FUNC_PARAMS,
     int64_t cuda_stream) {
   PrefillPlanSM90Info plan_info;
   plan_info.FromVector(plan_info_vec);
@@ -160,7 +160,7 @@ void BatchPrefillWithPagedKVCacheSM90Run(
     at::Tensor paged_v_cache, at::Tensor qo_indptr, at::Tensor paged_kv_indptr,
     at::Tensor paged_kv_indices, at::Tensor paged_kv_last_page_len, at::Tensor o,
     std::optional<at::Tensor> maybe_lse, int64_t mask_mode_code, int64_t layout,
-    int32_t window_left ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream) {
+    int64_t window_left ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream) {
   PrefillPlanSM90Info plan_info;
   plan_info.FromVector(plan_info_vec);
