+++ csrc/flashinfer_ops.cu
@@ -37,14 +37,14 @@ void merge_states(at::Tensor v, at::Tensor s, at::Tensor v_merged, at::Tensor s_
 
 void single_decode_with_kv_cache(at::Tensor q, at::Tensor k, at::Tensor v, at::Tensor tmp,
                                  at::Tensor o, int64_t layout,
-                                 int window_left SINGLE_DECODE_ADDITIONAL_FUNC_PARAMS,
+                                 int64_t window_left SINGLE_DECODE_ADDITIONAL_FUNC_PARAMS,
                                  int64_t cuda_stream);
 
 at::Tensor BatchDecodeWithPagedKVCachePlan(
     at::Tensor float_workspace_buffer, at::Tensor int_workspace_buffer,
     at::Tensor page_locked_int_workspace_buffer, at::Tensor indptr, int64_t batch_size,
     int64_t num_qo_heads, int64_t num_kv_heads, int64_t page_size,
-    bool enable_cuda_graph, int window_left, double logits_soft_cap, int64_t head_dim,
+    bool enable_cuda_graph, int64_t window_left, double logits_soft_cap, int64_t head_dim,
     at::Tensor empty_q_data, at::Tensor empty_kv_data, int64_t cuda_stream);
 
 void BatchDecodeWithPagedKVCacheRun(
@@ -52,7 +52,7 @@ void BatchDecodeWithPagedKVCacheRun(
     at::Tensor plan_info_vec, at::Tensor q, at::Tensor paged_k_cache,
     at::Tensor paged_v_cache, at::Tensor paged_kv_indptr, at::Tensor paged_kv_indices,
     at::Tensor paged_kv_last_page_len, at::Tensor o, std::optional<at::Tensor> maybe_lse,
-    int64_t kv_layout_code, int window_left BATCH_DECODE_ADDITIONAL_FUNC_PARAMS,
+    int64_t kv_layout_code, int64_t window_left BATCH_DECODE_ADDITIONAL_FUNC_PARAMS,
     int64_t cuda_stream);
 
 //========== gemm ==========
@@ -99,7 +99,7 @@ void block_sparse_indices_to_vector_sparse_offsets(at::Tensor block_sparse_indic
 void single_prefill_with_kv_cache(at::Tensor q, at::Tensor k, at::Tensor v, at::Tensor tmp,
                                   at::Tensor o, std::optional<at::Tensor> maybe_lse,
                                   int64_t mask_mode_code, int64_t layout,
-                                  int32_t window_left SINGLE_PREFILL_ADDITIONAL_FUNC_PARAMS,
+                                  int64_t window_left SINGLE_PREFILL_ADDITIONAL_FUNC_PARAMS,
                                   int64_t cuda_stream);
 
 at::Tensor BatchPrefillWithKVCachePlan(
@@ -114,7 +114,7 @@ void BatchPrefillWithRaggedKVCacheRun(
     at::Tensor plan_info_vec, at::Tensor q, at::Tensor k, at::Tensor v,
     at::Tensor qo_indptr, at::Tensor kv_indptr, at::Tensor o, std::optional<at::Tensor> maybe_lse,
     int64_t mask_mode_code, int64_t layout,
-    int32_t window_left BATCH_PREFILL_ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream);
+    int64_t window_left BATCH_PREFILL_ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream);
 
 void BatchPrefillWithPagedKVCacheRun(
     at::Tensor float_workspace_buffer, at::Tensor int_workspace_buffer,
@@ -122,7 +122,7 @@ void BatchPrefillWithPagedKVCacheRun(
     at::Tensor paged_v_cache, at::Tensor qo_indptr, at::Tensor paged_kv_indptr,
     at::Tensor paged_kv_indices, at::Tensor paged_kv_last_page_len, at::Tensor o,
     std::optional<at::Tensor> maybe_lse, int64_t mask_mode_code, int64_t layout,
-    int32_t window_left BATCH_PREFILL_ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream);
+    int64_t window_left BATCH_PREFILL_ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream);
 
 //========== quantization ==========
 
@@ -197,7 +197,7 @@ void chain_speculative_sampling(at::Tensor draft_probs, at::Tensor draft_token_i
                                 at::Tensor output_emitted_token_num, bool deterministic,
                                 int64_t cuda_stream);
 
-//========== pybind11 ==========
+//========== Torch Library ==========
 
 TORCH_LIBRARY(TORCH_EXTENSION_NAME, m) {
   // activation
